requirements.txt
Flask
SpeechRecognition
firebase-admin
langchain
langchain-openai
python-dotenv
openai
